Fix output branch selection
This fix is needed to address an issue that happens when running a chain of modules in the postprocessor with an outputbranchselection. Only the first module that is run has access to the full event, the second module has access only to the branches that are in the outputbranchselection, which is not the intended behaviour I believe. All modules should have access to all branches, and output branch selection should only affect which branches are written out.